Segurancao de acesso adicionada

+++ administracao/views.py
@@ -24,21 +24,21 @@ def index(request):
 
 
 def gerenciar_niveis(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_superuser:
         niveis = Nivel.objects.all()
         context = {"niveis": niveis}
         return render(request, "administracao/public/gerenciar-niveis.html", context)
+    return redirect("/")
 
 
 def novo_nivel(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_superuser:
         form = FormNovoNivel(request.POST or None)
         if request.method == "POST":
             if form.is_valid():
                 nivel = request.POST.get("nome_do_nivel")
                 Nivel.objects.create(nome_do_nivel=nivel)
-                messages.add_message(
-                    request, messages.SUCCESS, "Nivel cadastrado com sucesso!")
+                messages.add_message(request, messages.SUCCESS, "Nivel cadastrado com sucesso!")
                 return redirect(f"/administracao/niveis/")
         context = {"form": form}
         return render(request, "administracao/public/novo-nivel.html", context)
@@ -46,11 +46,12 @@ def novo_nivel(request):
 
 
 def excluir_nivel(request, nivel):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_superuser:
         Nivel.objects.get(id=nivel).delete()
         messages.add_message(
             request, messages.SUCCESS, "Nivel removido com sucesso!")
         return redirect(f"/administracao/niveis/")
+    return redirect(f"/")
 
 ##########################################################
 ############## GERENCIAMENTO DE USUARIOS #################
@@ -68,35 +69,46 @@ def gerenciar_usuarios(request, slug):
 
 def novo_usuario(request, slug):
     if request.user.is_authenticated:
-        form = FormNovoUsuario(request.POST or None)
-        if request.method == "POST":
-            if form.is_valid():
-                usuario = request.POST.get("usuario")
-                senha = request.POST.get("senha")
-                email = request.POST.get("email")
-                nivel = request.POST.get("nivel")
+        try:
+            nivel = Login.objects.get(usuario=request.user.id)
+        except:
+            pass
+        if request.user.is_superuser or nivel.nivel.id == 1:
+            form = FormNovoUsuario(request.POST or None)
+            if request.method == "POST":
+                if form.is_valid():
+                    usuario = request.POST.get("usuario")
+                    senha = request.POST.get("senha")
+                    email = request.POST.get("email")
+                    nivel = request.POST.get("nivel")
 
-                user = User.objects.create_user(
-                    username=usuario, password=senha, email=email)
-                emp = Empresa.objects.get(slug_da_empresa=slug)
+                    user = User.objects.create_user(
+                        username=usuario, password=senha, email=email)
+                    emp = Empresa.objects.get(slug_da_empresa=slug)
 
-                nivel = Nivel.objects.get(id=nivel)
-                Login.objects.create(usuario=user, empresa=emp, nivel=nivel)
+                    nivel = Nivel.objects.get(id=nivel)
+                    Login.objects.create(usuario=user, empresa=emp, nivel=nivel)
 
-                messages.add_message(
-                    request, messages.SUCCESS, "Usuario cadastrado com sucesso!")
-                return redirect(f"/administracao/usuarios/{slug}/")
-        context = {"form": form, "slug": slug}
-        return render(request, "administracao/public/novo-usuario.html", context)
+                    messages.add_message(
+                        request, messages.SUCCESS, "Usuario cadastrado com sucesso!")
+                    return redirect(f"/administracao/usuarios/{slug}/")
+            context = {"form": form, "slug": slug}
+            return render(request, "administracao/public/novo-usuario.html", context)
     return redirect("/")
 
 
 def remove_usuario(request, slug, usuario):
     if request.user.is_authenticated:
-        User.objects.get(id=usuario).delete()
-        messages.add_message(
-            request, messages.SUCCESS, "Usuario removido com sucesso!")
-        return redirect(f"/administracao/usuarios/{slug}/")
+        try:
+            nivel = Login.objects.get(usuario=request.user.id)
+        except:
+            pass
+        if request.user.is_superuser or nivel.nivel.id == 1:
+            User.objects.get(id=usuario).delete()
+            messages.add_message(
+                request, messages.SUCCESS, "Usuario removido com sucesso!")
+            return redirect(f"/administracao/usuarios/{slug}/")
+    return redirect(f"/")
 
 ##########################################################
 ############# GERENCIAMENTO DE PROJETOS ##################
@@ -105,42 +117,58 @@ def remove_usuario(request, slug, usuario):
 
 def gerenciar_projetos(request, slug):
     if request.user.is_authenticated:
-        projetos = Projeto.objects.filter(
-            empresa=Empresa.objects.get(slug_da_empresa=slug).id)
-        context = {"slug": slug, "projetos": projetos}
-        return render(request, "administracao/public/gerenciar-projetos.html", context)
+        try:
+            nivel = Login.objects.get(usuario=request.user.id)
+        except:
+            pass
+        if request.user.is_superuser or nivel.nivel.id == 1:
+            projetos = Projeto.objects.filter(
+                empresa=Empresa.objects.get(slug_da_empresa=slug).id)
+            context = {"slug": slug, "projetos": projetos}
+            return render(request, "administracao/public/gerenciar-projetos.html", context)
     return redirect("/")
 
 
 def novo_projeto(request, slug):
     if request.user.is_authenticated:
-        form = FormNovoProjeto(request.POST or None)
-        if request.method == "POST":
-            if form.is_valid():
-                nome_do_projeto = request.POST.get("nome_do_projeto")
-                slug_do_projeto = nome_do_projeto.replace(" ", "-")
-                emp = Empresa.objects.get(slug_da_empresa=slug)
-                Projeto.objects.create(
-                    slug_do_projeto=slug_do_projeto, nome_do_projeto=nome_do_projeto, empresa=emp)
-                cria_pasta(f"media/{slug}/{slug_do_projeto}")
-                cria_pasta_arquivos(slug=slug, projeto=slug_do_projeto)
-                messages.add_message(
-                    request, messages.SUCCESS, "Projeto cadastrado com sucesso!")
-                return redirect(f"/administracao/projetos/{slug}/")
-        context = {"form": form, "slug": slug}
-        return render(request, "administracao/public/novo-projeto.html", context)
+        try:
+            nivel = Login.objects.get(usuario=request.user.id)
+        except:
+            pass
+        if request.user.is_superuser or nivel.nivel.id == 1:
+            form = FormNovoProjeto(request.POST or None)
+            if request.method == "POST":
+                if form.is_valid():
+                    nome_do_projeto = request.POST.get("nome_do_projeto")
+                    slug_do_projeto = nome_do_projeto.replace(" ", "-")
+                    emp = Empresa.objects.get(slug_da_empresa=slug)
+                    Projeto.objects.create(
+                        slug_do_projeto=slug_do_projeto, nome_do_projeto=nome_do_projeto, empresa=emp)
+                    cria_pasta(f"media/{slug}/{slug_do_projeto}")
+                    cria_pasta_arquivos(slug=slug, projeto=slug_do_projeto)
+                    messages.add_message(
+                        request, messages.SUCCESS, "Projeto cadastrado com sucesso!")
+                    return redirect(f"/administracao/projetos/{slug}/")
+            context = {"form": form, "slug": slug}
+            return render(request, "administracao/public/novo-projeto.html", context)
     return redirect("/")
 
 
 def remove_projeto(request, slug, projeto):
     if request.user.is_authenticated:
-        projeto = Projeto.objects.get(id=projeto)
-        shutil.rmtree(f"{settings.BASE_DIR}/media/{slug}/{projeto.slug_do_projeto}",
-                      ignore_errors=True)
-        projeto.delete()
-        messages.add_message(
-            request, messages.SUCCESS, "Projeto deletado com sucesso!")
-        return redirect(f"/administracao/projetos/{slug}/")
+        try:
+            nivel = Login.objects.get(usuario=request.user.id)
+        except:
+            pass
+        if request.user.is_superuser or nivel.nivel.id == 1:
+            projeto = Projeto.objects.get(id=projeto)
+            shutil.rmtree(f"{settings.BASE_DIR}/media/{slug}/{projeto.slug_do_projeto}",
+                        ignore_errors=True)
+            projeto.delete()
+            messages.add_message(
+                request, messages.SUCCESS, "Projeto deletado com sucesso!")
+            return redirect(f"/administracao/projetos/{slug}/")
+    return redirect("/")
 
 ##########################################################
 ############# GERENCIAMENTO DE ARQUIVOS ##################
